+++ src/gauss_seidel.h
@@ -1,4 +1,4 @@
 #include <vector>
 #include "utils.h"
 
-void run_gs(const matrix& A, const vector<float>& b, vector<float>& x, int max_iter = 1000, float tol = 1e-6, float x_0 = 0.0);
+void run_gs(const matrix& A, const vector<double>& b, vector<double>& x, int max_iter = 1000, double tol = 1e-6, double x_0 = 0.0);
