+++ src/gmres.h
@@ -3,8 +3,8 @@
 
 void run_gmres(
     const matrix& A, 
-    const vector<float>& b, 
-    vector<float>& x, 
+    const vector<double>& b, 
+    vector<double>& x, 
     int max_iter = 1000, 
-    float tol = 1e-10
+    double tol = 1e-10
 );
