+++ src/jacobi.cpp
@@ -4,20 +4,20 @@
 
 void run_jacobi(
     const matrix& A, 
-    const vector<float>& b, 
-    vector<float>& x, 
+    const vector<double>& b, 
+    vector<double>& x, 
     int max_iter, 
-    float tol, 
-    float x_0
+    double tol, 
+    double x_0
 ) {
     int n = A.size();
-    vector<float> x_old(n, x_0);  
-    vector<float> x_new(n, 0.0); 
-    float error;
+    vector<double> x_old(n, x_0);  
+    vector<double> x_new(n, 0.0); 
+    double error;
 
     for (int iter = 0; iter < max_iter; iter++) {
         for (int i = 0; i < n; i++) {
-            float sum = b[i];
+            double sum = b[i];
             for (int j = 0; j < n; j++) {
                 if (i != j) {
                     sum -= A[i][j] * x_old[j];
