+++ src/utils.cpp
@@ -1,7 +1,7 @@
 #include "utils.h"
 
-float row_sum(int i, const matrix& A) {
-    float sum = 0;
+double row_sum(int i, const matrix& A) {
+    double sum = 0;
     for (int j = 0; j < A.size(); j++) {  
         if (i != j) { sum += fabs(A[i][j]); }
     }
@@ -17,9 +17,9 @@ bool is_gs_j_convergent(const matrix& A) {
 }
 
 
-vector<float> matrix_vec_mult(const matrix& A, const vector<float>& v) {
+vector<double> matrix_vec_mult(const matrix& A, const vector<double>& v) {
     int n = A.size();
-    vector<float> result(n, 0.0);
+    vector<double> result(n, 0.0);
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
             result[i] += A[i][j] * v[j];
@@ -28,32 +28,32 @@ vector<float> matrix_vec_mult(const matrix& A, const vector<float>& v) {
     return result;
 }
 
-float dot_product(const vector<float>& a, const vector<float>& b) {
-    float result = 0.0;
+double dot_product(const vector<double>& a, const vector<double>& b) {
+    double result = 0.0;
     for (int i = 0; i < a.size(); i++) {
         result += a[i] * b[i];
     }
     return result;
 }
 
-vector<float> scale_vector(float alpha, const vector<float>& v) {
-    vector<float> result(v.size(), 0.0);
+vector<double> scale_vector(double alpha, const vector<double>& v) {
+    vector<double> result(v.size(), 0.0);
     for (int i = 0; i < v.size(); i++) {
         result[i] = alpha * v[i];
     }
     return result;
 }
 
-vector<float> subtract_vectors(const vector<float>& a, const vector<float>& b) {
-    vector<float> result(a.size(), 0.0);
+vector<double> subtract_vectors(const vector<double>& a, const vector<double>& b) {
+    vector<double> result(a.size(), 0.0);
     for (int i = 0; i < a.size(); i++) {
         result[i] = a[i] - b[i];
     }
     return result;
 }
 
-void normalize_vector(vector<float>& v) {
-    float norm = 0.0;
+void normalize_vector(vector<double>& v) {
+    double norm = 0.0;
     for (int i = 0; i < v.size(); i++) {
         norm += v[i] * v[i];
     }
@@ -63,8 +63,8 @@ void normalize_vector(vector<float>& v) {
     }
 }
 
-float vector_norm(const vector<float>& v) {
-    float norm = 0.0;
+double vector_norm(const vector<double>& v) {
+    double norm = 0.0;
     for (int i = 0; i < v.size(); i++) {
         norm += v[i] * v[i];
     }
