+++ src/utils.h
@@ -2,13 +2,13 @@
 #include <cmath>
 using std::vector;
 
-typedef vector<vector<float>> matrix;
+typedef vector<vector<double>> matrix;
 
-float row_sum(int i, const matrix& A);
+double row_sum(int i, const matrix& A);
 bool is_gs_j_convergent(const matrix& A);
-vector<float> matrix_vec_mult(const matrix& A, const vector<float>& v);
-float dot_product(const vector<float>& a, const vector<float>& b);
-vector<float> scale_vector(float alpha, const vector<float>& v);
-vector<float> subtract_vectors(const vector<float>& a, const vector<float>& b);
-void normalize_vector(vector<vector<float>>& v);
-float vector_norm(const vector<float>& v);
+vector<double> matrix_vec_mult(const matrix& A, const vector<double>& v);
+double dot_product(const vector<double>& a, const vector<double>& b);
+vector<double> scale_vector(double alpha, const vector<double>& v);
+vector<double> subtract_vectors(const vector<double>& a, const vector<double>& b);
+void normalize_vector(vector<vector<double>>& v);
+double vector_norm(const vector<double>& v);
